Add basic __vectorcall support
This is a fake implementation, which piggy-backs on the x64 standard
calling convention and __fastcall, for x86_64 and x86 respectively,
without any modifications.  It won't work for some vector type
parameters and for any HVA types.  Since SIMD vector types and HVA are
not supported by ext/ffi anyway, only float/double parameters would be
a real issue.